Replace default render impl with single template
