+++ adhocracy4/emails/mixins.py
@@ -1,25 +1,6 @@
 from email.mime.image import MIMEImage
 
 from django.contrib.staticfiles import finders
-from django.utils.translation import get_language
-from django.template.loader import select_template
-
-
-class SingleTemplateMixin():
-    def render(self, template_name, context):
-        languages = [get_language(), self.fallback_language]
-        template = select_template([
-            'emails/{}.{}.email'.format(template_name, lang)
-            for lang in languages
-        ])
-
-        parts = []
-        for part_type in ('subject', 'txt', 'html'):
-            context.update({'part_type': part_type})
-            parts.append(template.render(context))
-            context.pop()
-
-        return tuple(parts)
 
 
 class PlatformEmailMixin():
